added `xml2png_5` and `xml2png_spectra` to `scripts`
