+++ examples/xml2png_5/README.md
@@ -1,0 +1,6 @@
+# Example illustrating the use of xml2png script
+
+```bash
+xml2png_5 *.xml
+```
+
